requirements.txt
discord
python-dotenv
pymongo
google
dnspython